+++ example-code/app/src/backend/test/KnownTestLocations.ts
@@ -9,12 +9,6 @@ export class KnownTestLocations {
 
   /** The directory where test assets are stored. Keep in mind that the test is playing the role of the app. */
   public static get assetsDir(): string {
-    const imodeljsMobile = Platform.imodeljsMobile;
-    if (imodeljsMobile !== undefined) {
-      return imodeljsMobile.knownLocations.assetsDir;
-    }
-
-    // Assume that we are running in nodejs
     return path.join(__dirname, "assets");
   }
 
@@ -28,5 +22,4 @@ export class KnownTestLocations {
     // Assume that we are running in nodejs
     return path.join(__dirname, "output");
   }
-
 }